stdio.c: fix compiler warning

+++ firmware/libcommon/source/stdio.c
@@ -294,7 +294,7 @@ signed int vsnprintf(char *pStr, size_t length, const char *pFormat, va_list ap)
     char          fill;
     unsigned char width;
     signed int    num = 0;
-    signed int    size = 0;
+    size_t        size = 0;
 
     // Clear the string
     if (pStr) {
